commit for pull

+++ src/main/java/com/wor/dash/record/model/Record.java
@@ -15,12 +15,12 @@ public class Record {
 
     private static final DateTimeFormatter timeFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd hh:mm:ss a");
     private Integer userId;
-    private Integer totalCalorie; // 총소모칼로리
+    private Double totalCalorie; // 총소모칼로리
     private String startTime; // 운동 시작 시간
     private Integer meanHeartRate; // 평균 심박수
     private Integer count; // 걸음수
     private Integer maxHeartRate; // 최대 심박수
-    private Integer maxSpeed; // 최대 속도 (m/s)
+    private Double maxSpeed; // 최대 속도 (m/s)
     private Integer minHeartRate; // 최소 심박수
     private Double distance; // 이동 거리(미터)
     private Double meanSpeed;
